requirements.txt
https://www.djangoproject.com/download/1.7c1/tarball/
Fabric==1.9.0
Markdown==2.4.1
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
blist==1.3.6
boto==2.31.1
cassandra-driver==1.0.2
celery==3.1.13
dj-database-url==0.3.0
dj-static==0.0.6
django-celery==3.1.10
django-extensions==1.3.8
django-filter==0.7
django-grappelli==2.4.7
django-nose==1.2
django-rest-framework-docs==0.1.7
django-rest-swagger==0.1.9
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.3.14
ecdsa==0.11
feedly==0.10.10
futures==2.1.6
gunicorn==19.1.0
jsonpickle==0.7.1
kombu==3.0.21
meld3==1.0.0
nose==1.3.3
paramiko==1.14.0
psycopg2==2.5.1
pycrypto==2.6.1
pyparsing==1.5.7
pystache==0.5.4
pytz==2014.4
redis==2.10.1
scales==1.0.6
simplejson==3.6.0
six==1.7.3
static==1.0.2
static3==0.5.1
supervisor==3.0
wsgiref==0.1.2
